+++ CHANGELOG.md
@@ -5,6 +5,33 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 ## [Unreleased]
 
+## [8.33.0]
+
+### Added
+
+- managed object storage: add support for policy versions
+
+### Fixed
+
+- account: improve parsing of `Firewall` field in `BillingResourceDetail` to handle both string and float64 types
+
+## [8.32.0]
+
+### Removed
+
+- file-storage: remove `Shares` from `ModifyFileStorageRequest` and `ReplaceFileStorageRequest`
+- file-storage: remove `Path` from `ModifyFileStorageShareRequest`
+
+### Added
+
+- file-storage: add File Storage Share ACL management support
+- file-storage: add `Deleting` boolean to `FileStorageShare` in upcloud package
+- file-storage: add `FileStorageShare` to request package
+- file-storage: add `GetFileStorageCurrentState` method for getting the effective state of the service
+
+### Changed
+- file-storage: rename `FileStorageACL` to `FileStorageShareACL` and related constants accordingly in upcloud package
+
 ## [8.31.1]
 
 ### Changed
@@ -694,7 +721,9 @@ See updating [Changelog example here](https://keepachangelog.com/en/1.0.0/)
 
 First stable release
 
-[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.31.1...HEAD
+[Unreleased]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.33.0...HEAD
+[8.33.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.32.0...v8.33.0
+[8.32.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.31.1...v8.32.0
 [8.31.1]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.31.0...v8.31.1
 [8.31.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.30.0...v8.31.0
 [8.30.0]: https://github.com/UpCloudLtd/upcloud-go-api/compare/v8.29.2...v8.30.0